Move functions to circuit_frame
